refactor: rename hoc
